requirements.txt
# Variáveis de ambiente
python-dotenv==1.0.1

# Hugging Face
transformers==4.41.2
torch==2.3.0
huggingface_hub==0.23.3

# LangChain + FAISS + Embeddings locais
langchain==0.2.1
langchain-community==0.2.1
sentence-transformers==2.7.0
faiss-cpu==1.7.4

# Leitura de arquivos
unstructured
pdfminer.six==20221105
docx2txt==0.8

# Gradio para interface
gradio==4.34.0

# Extra dependencies
scikit-learn